Fix openssl build
This commit adds changes to specify debian suffix so we can pull latest stable version of openssl available from upstream. Earlier specified version is not available on upstream mirrors and going through history - it was bumped to fix a CVE which has been fixed in the version specified in this commit (https://www.debian.org/security/2023/dsa-5532.en.html).